Carry ClickHouse counter scans as doubles
Read ClickHouse sums as doubles in both metrics and PromQL adapters, preserve fractional cumulative-to-delta conversion, and migrate SumScan away from integer counts. Leave evaluator-wide gauge precision migration for the dedicated cleanup pass.

+++ okapi-ingester/src/main/java/org/okapi/metrics/ch/ChSumSample.java
@@ -7,4 +7,9 @@ package org.okapi.metrics.ch;
 import java.util.Map;
 
 public record ChSumSample(
-    long tsStart, long tsEnd, long value, CH_SUM_TYPE sumType, Map<String, String> tags, String unit) {}
+    long tsStart,
+    long tsEnd,
+    double value,
+    CH_SUM_TYPE sumType,
+    Map<String, String> tags,
+    String unit) {}
